[libc++] Extract a clean base support API for std::atomic
This doesn't change the operations that std::atomic is based on,
but it reorganizes the C11 / GCC implementation split to make it
clearer what's the base support layer and what's not.
